third_party/WebKit/LayoutTests/fast/scrolling/absolute-position-overflow-clip.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body id="body">
    <div style="height: 500px;overflow-x: hidden;">
        <div style="height: 2000px; background-color: blue;"></div>
        <div style="position: absolute;">
            <div style="height: 1000px; background-color: yellow;"></div>
            <a href="somelink" id="focusable" style="font-size:60px">anchor</a>
        </div>
        <div style="height: 2000px; background-color: red;"></div>
    </div>
</body>
<script type="text/javascript">
window.onload = function() {
    test(function(t) {
        var focusable = document.getElementById("focusable");
        focusable.focus();
        assert_not_equals(window.scrollY, 0);
    }, "Tests that for an element with absolute position and whose parent has overflowclip, it can still scroll automatically.");
}
</script>
